made the suggested changes
