+++ src/view.rs
@@ -2,6 +2,7 @@ use std::fmt::Display;
 
 pub mod editor;
 pub mod terminal;
+pub mod configuration;
 
 pub use editor::NanoEditor;
 pub use terminal::{TerminalSize, TerminalView};
